pulled out trim extensions util to use in cloudinary and synced on filenameAsPublicID option with UID default
